Bug Fixes by Chromabit (#11)
* Make the cooldown only apply to turning off PVP

* Make blocked worlds disable PVP

Previously the blocked worlds settings just stopped processing of the events by the plugin; this makes it so the events get cancelled.

* Allow for PVP to be completely enabled/disabled per world

If a world has PVP off, every player in that world will be set to have PVP disabled. If a world has PVP on and it is in the blocked worlds, every player in that world will be set to have PVP enabled.